chore: address issue #614
