+++ app/services/refraction_static_artifacts/__init__.py
@@ -4,6 +4,8 @@ from __future__ import annotations
 
 from app.services.refraction_static_artifacts import contract
 from app.services.refraction_static_artifacts import _legacy
+from app.services.refraction_static_artifacts import first_break_qc
+from app.services.refraction_static_artifacts import reduced_time_qc
 from app.services.refraction_static_artifacts import solution
 from app.services.refraction_static_artifacts._legacy import *  # noqa: F403
 
@@ -19,6 +21,11 @@ TIME_TERM_SPREADSHEET_FORMAT_VERSION = contract.TIME_TERM_SPREADSHEET_FORMAT_VER
 TIME_TERM_SPREADSHEET_SCHEMA_VERSION = contract.TIME_TERM_SPREADSHEET_SCHEMA_VERSION
 
 __all__ = list(_legacy.__all__)
+for _module in (first_break_qc, reduced_time_qc):
+    for _name in _module.__all__:
+        globals()[_name] = getattr(_module, _name)
+        if _name not in __all__:
+            __all__.append(_name)
 for _name in solution.__all__:
     globals()[_name] = getattr(solution, _name)
     if _name not in __all__:
